Basic three-phase setup + RFB security type probe
Added basic support for a three-phase setup activity, although the first
and second phase are currently identical.  The first phase is just for
the bare essentials (hostname and port) needed to probe the VNC (RFB)
server to query its version and supported security types.  The second
phase collects more information from the user based on the results of
the probe.

The eventual plan:

Phase 1:
- Collect IP/hostname and port from the user.
- Connect to the server to probe the version number and security types.

Phase 2:
- Collect extra information, such as the name to be used in the
  selection list.
- If phase 1 detected an Apple Remote Desktop (ARD) server, offer to
  enable ARD v3.5 compatibility (send button-2 instead of button-3
  on two-finger-tap).
- If phase 1 detected the availability of ARD authentication, offer to
  enable "Mac authentication" with a username.
- Connect to the server to test the authentication credentials.

Phase 3:
- Show the final results and ask for confirmation.
- Add the server to the persistent storage.

Next steps:

- detect ARD auth
- implement ARD auth