Adding Mac authentication support for Apple Remote Desktop.
The setup page has been reworked to accommodate this new authentication
scheme, and the RFB authentication model has been made extensible to
allow for swappable authentication methods.  Mac authentication is
required to use Valence with Mac OS X 10.7 Lion.

The RFBSecurityARD class implements "Mac Authentication", which uses
Diffie-Hellman key agreement (along with MD5 and AES128) to authenticate
users to Apple Remote Desktop, the VNC server which is built-in to
Mac OS X.

This authentication technique is based on the following steps:

1. Perform Diffie-Hellman key agreement, so both sides have
   a shared secret key which can be used for further encryption.
2. Take the MD5 hash of this DH secret key to produce a 128-bit
   value which we will use as the actual encryption key.
3. Encrypt the username and password with this key using the AES
   128-bit symmetric cipher in electronic codebook (ECB) mode.  The
   username/password credentials are stored in a 128-byte structure,
   with 64 bytes for each, null-terminated.  Ideally, write random
   values into the portion of this 128-byte structure which is not
   occupied by the username or password, but no further padding for
   this block cipher.

The ciphertext from step 3 and the DH public key from step 2
are sent to the server.